+++ .eslintrc.json
@@ -23,6 +23,7 @@
       "no-underscore-dangle": "off",
       "max-len": "off",
       "import/no-anonymous-default-export": "off",
-      "no-case-declarations": "off"
+      "no-case-declarations": "off",
+      "consistent-return": "off"
     }
 }
